+++ include/game/item.h
@@ -15,19 +15,23 @@ protected:
     bool pickedUp = false;
 
     void setTexture(std::string texture);
+    void updateDropVelocity();
+    void updatePickedUpPosition();
+    void updateCollisionWithPlayer();
+    virtual void updateImpl() = 0;
 
 public:
     Item(Game* game, LevelPos pos, std::string texture);
-    virtual ~Item() = 0;
+    virtual ~Item();
 
     void update();
     void render();
 
     void pickUp();
+    virtual void collides() = 0;
 
     LevelPos getPos() const;
     LevelVelocity getVelocity() const;
-    bool isPickedUp() const;
     std::string getTexture() const;
 };
 
@@ -37,11 +41,16 @@ class Xp : public Item {
 private:
     int level;
 
+    void updateMerge();
+
 public:
     Xp(Game* game, LevelPos pos, int level);
     ~Xp() override;
 
     void increaseLevel();
+    void collides() override;
+
+    void updateImpl() override;
 
     int getLevel() const;
 };
@@ -54,6 +63,10 @@ private:
 public:
     Coin(Game* game, LevelPos pos);
     ~Coin() override;
+
+    void collides() override;
+
+    void updateImpl() override;
 };
 
 #endif