+++ src/coin.cpp
@@ -9,3 +9,15 @@ Coin::Coin(Game* g, LevelPos p) : Item(g, p, "coin.png") {
 Coin::~Coin() {
     
 }
+
+// EVENTS
+void Coin::collides() {
+    Map* map = game->getMap();
+    map->removeItem(this);
+    map->getPlayer()->increaseCoins(5);
+}
+
+// UPDATE
+void Coin::updateImpl() {
+    
+}