+++ src/enemy.cpp
@@ -18,10 +18,14 @@ bool Enemy::isNearPlayer() {
 }
 
 void Enemy::kill() {
+    static float enemyDropingCoinProbability = game->getData()->probabilities->enemyDropingCoin;
     Map* map = game->getMap();
 
     map->dropXp(pos, data.xpLevel);
     map->killEnemy(this);
+    map->getPlayer()->increaseCoins(1);
+
+    if (rand(enemyDropingCoinProbability)) map->dropCoin(pos);
 }
 
 // UPDATE
@@ -37,7 +41,7 @@ void Enemy::updateForPlayer() {
         LevelPos playerPos = map->getPlayer()->getPos();
 
         float angle = std::atan2(playerPos.y - pos.y, playerPos.x - pos.x);
-        LevelVelocity velocity = { data.stats.speed * std::cos(angle), data.stats.speed * std::sin(angle) };
+        velocity = { data.stats.speed * std::cos(angle), data.stats.speed * std::sin(angle) };
         pos = pos + velocity;
 
         if (velocity.x != 0) facing = (velocity.x > 0) ? RIGHT : LEFT;