+++ src/item.cpp
@@ -1,7 +1,7 @@
 #include "item.h"
 
 // INIT
-Item::Item(Game* g, LevelPos p, std::string t) : game(g), pos(p), sprite(new Sprite(g, t, 200)), texture(t) {
+Item::Item(Game* g, LevelPos p, std::string t) : game(g), pos(p), sprite(new Sprite(g, t, 150)), texture(t) {
     float angle = rand(0.0f, M_PI * 2.0f);
     velocity = { std::cos(angle), std::sin(angle) };
 }
@@ -18,31 +18,52 @@ void Item::pickUp() {
 
 // UPDATE
 void Item::update() {
-    static float airResistance = game->getData()->physics->airResistance;
-    static float itemSpeed = game->getData()->physics->itemSpeed;
-
     sprite->update();
 
-    if (!pickedUp) {
-        if (velocity.isNull()) return;
+    if (pickedUp) {
+        updatePickedUpPosition();
+        updateCollisionWithPlayer();
+    } else updateDropVelocity();
 
-        if (velocity.x != 0.0f) {
-            if (velocity.x < 0.0f) velocity.x = std::min(0.0f, velocity.x + airResistance);
-            else velocity.x = std::max(0.0f, velocity.x - airResistance);
-        }
-        if (velocity.y != 0.0f) {
-            if (velocity.y < 0.0f) velocity.y = std::min(0.0f, velocity.y + airResistance);
-            else velocity.y = std::max(0.0f, velocity.y - airResistance);
-        }
+    updateImpl();
+}
 
-        pos = pos + velocity * (itemSpeed / 2);
-    } else {
-        LevelPos playerPos = game->getMap()->getPlayer()->getPos();
+void Item::updateDropVelocity() {
+    static GameData* gameData = game->getData();
+    static float airResistance = gameData->physics->airResistance;
+    static float itemSpeed = gameData->physics->itemSpeed;
 
-        float angle = std::atan2(playerPos.y - pos.y, playerPos.x - pos.x);
-        velocity = { std::cos(angle), std::sin(angle) };
-        pos = pos + velocity * itemSpeed;
+    if (velocity.isNull()) return;
+
+    if (velocity.x != 0.0f) {
+        if (velocity.x < 0.0f) velocity.x = std::min(0.0f, velocity.x + airResistance);
+        else velocity.x = std::max(0.0f, velocity.x - airResistance);
     }
+    if (velocity.y != 0.0f) {
+        if (velocity.y < 0.0f) velocity.y = std::min(0.0f, velocity.y + airResistance);
+        else velocity.y = std::max(0.0f, velocity.y - airResistance);
+    }
+
+    pos = pos + velocity * (itemSpeed / 2);
+}
+
+void Item::updatePickedUpPosition() {
+    static float itemSpeed = game->getData()->physics->itemSpeed;
+
+    LevelPos playerPos = game->getMap()->getPlayer()->getPos();
+
+    float angle = std::atan2(playerPos.y - pos.y, playerPos.x - pos.x);
+    velocity = { std::cos(angle), std::sin(angle) };
+    pos = pos + velocity * itemSpeed;
+}
+
+void Item::updateCollisionWithPlayer() {
+    static std::unordered_map<std::string, LevelRect> itemHitboxes = game->getData()->physics->itemHitboxes;
+    static LevelRect defaultPlayerHitbox = game->getData()->others->entities[0].hitbox;
+
+    LevelRect playerHitbox = defaultPlayerHitbox + game->getMap()->getPlayer()->getPos();
+    LevelRect itemHitbox = itemHitboxes[texture] + pos;
+    if (playerHitbox.collides(itemHitbox) != NONE) collides();
 }
 
 // RENDER
@@ -54,11 +75,11 @@ void Item::render() {
 
     static std::unordered_map<std::string, LevelSize> texturesSizes = gameData->sizes->textures;
     static std::unordered_map<std::string, LevelPos> texturesOffsets = gameData->offsets->textures;
-    static std::unordered_map<std::string, LevelRect> itemHitobxes = gameData->physics->itemHitobxes;
+    static std::unordered_map<std::string, LevelRect> itemHitboxes = gameData->physics->itemHitboxes;
 
     LevelPos texturePos = pos + texturesOffsets[texture];
     LevelSize textureSize = texturesSizes[texture];
-    LevelRect defaultHitbox = itemHitobxes[texture];
+    LevelRect defaultHitbox = itemHitboxes[texture];
     LevelRect hitbox = defaultHitbox + pos;
 
     if (!camera->contains(texturePos.makeRect(textureSize))) return;
@@ -81,10 +102,6 @@ LevelPos Item::getPos() const {
 
 LevelVelocity Item::getVelocity() const {
     return velocity;
-}
-
-bool Item::isPickedUp() const {
-    return pickedUp;
 }
 
 std::string Item::getTexture() const {