+++ src/xp.cpp
@@ -19,6 +19,48 @@ void Xp::increaseLevel() {
     setTexture("xp-lvl-" + std::to_string(level) + ".png");
 }
 
+void Xp::collides() {
+    Map* map = game->getMap();
+    map->removeItem(this);
+    map->getPlayer()->increaseXp(level);
+}
+
+// UPDATE
+void Xp::updateImpl() {
+    updateMerge();
+}
+
+void Xp::updateMerge() {
+    static int maxXpLevel = game->getData()->others->maxXpLevel;
+    static std::unordered_map<std::string, LevelRect> itemHitboxes = game->getData()->physics->itemHitboxes;
+
+    Map* map = game->getMap();
+    if (map->hasToBeRemoved(this)) return;
+    LevelRect xpHitbox = itemHitboxes[texture] + pos;
+
+    for (Item* item : map->getItems()) {
+        Xp* other = dynamic_cast<Xp*>(item);
+        if (!other) continue; 
+
+        if (other == this) continue;
+        if (map->hasToBeRemoved(other)) continue;
+        if (level >= maxXpLevel) continue;
+        if (level != other->getLevel()) continue;
+
+        LevelRect otherHitbox = itemHitboxes[other->getTexture()] + other->getPos();
+        if (xpHitbox.collides(otherHitbox) == NONE) continue;
+
+        if (velocity > other->getVelocity()) {
+            increaseLevel();
+            map->removeItem(other);
+        } else {
+            other->increaseLevel();
+            map->removeItem(this);
+        }
+        break;
+    }
+}
+
 // GETTER
 int Xp::getLevel() const {
     return level;